+++ Utils/sql_queries/create_package_table.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS package(
    author         CHAR,
    license        CHAR,
    published_on   CHAR,
-   score          INT,
-   weigthed_score INT,
+   score          DECIMAL,
+   weigthed_score DECIMAL,
    decision       CHAR
 );
